.github/runners/generate/requirements.txt
pathlib
pyyaml